+++ src/hayate_fetch/__init__.py
@@ -8,7 +8,7 @@ from hayate import Request, Response
 
 from .backends import FetchBackend, UrllibBackend, WorkersBackend, default_backend
 
-__version__ = "0.1.3"
+__version__ = "0.1.4"
 
 __all__ = [
     "FetchBackend",